Changed layout of modal. Added Succcess component. Changed color styles of different modal view
