+++ src/assets/css/utilities.css
@@ -211,19 +211,23 @@
 }
 
 .border-success {
-    border: 1px solid var(--clr-success);
+    border: 2px solid var(--clr-success);
 }
 
 .border-blue {
-    border: 1px solid var(--clr-blue);
+    border: 2px solid var(--clr-blue);
 }
 
 .border-light {
-    border: 1px solid var(--clr-light);
+    border: 2px solid var(--clr-light);
 }
 
 .border-primary {
     border: 1px solid var(--clr-primary);
+}
+
+.border-red {
+    border: 2px solid var(--clr-red);
 }
 
 .border-none {
@@ -244,6 +248,14 @@
 
 .color-danger {
     color: var(--clr-danger);
+}
+
+.color-red {
+    color: var(--clr-red);
+}
+
+.color-blue {
+    color: var(--clr-blue);
 }
 
 .color-success {
